feat: add API endpoints for jobs
To be used initially as an internal API (for front ends that require it,
like a mobile app or other internal admin tools). It includes an initial
authentication of the client via deactivables API Keys.
Publishing jobs is optional at the time of job creation and the job can
be updated later to be published.
Deleting a job or adding a company is not supported by this initial set
of endpoints. Admins are required to to that via console or other
methods.