feat: unpublish jobs after a certain period
To prevent old jobs (that are probably already taken or not relevant
anymore) from appearing on the list, an solid queue job is scheduled to
unpublish the job after a configurable number of days after publishing.

Because unpublishing a job is a side effect that is not directly related
the job creation/publishing and it is not required to be synchronous to
them, a pub/sub architecture has been used for the implementation.

Additional tools have been added to the codebase in order for the
feature to be supported:

- active_event_store as the pub/sub library
- solid queue configurations and database setup (configuring also solid
  queue in development instead of the default in-memory jobs)
- mission_control-jobs to facilitate jobs management through a UI
  interface