feat: add pagination
Supported by both the API endpoints and web. We are using pagination by
offset (not by cursor) assuming the number of records is reasonable for
it.
In addition to page/per_page that can be passed as parameters (in web
always pages contain 10 elements), the rest of the parameters for each
page are kept. For example records will stay sorted or filtered even
when switching pages.

The job filtering code was extracted to a concern to keep related code
in the same file (no modification to behaviors), which is being tested
by shared examples (that the reason of the small configuration change i
tests).

_This is still a work in progress not ready for production. Additional
review/testing is required before shipping_