feat: add readme and final improvements
This commit adds the readme with the summary of the tasks implemented,
some examples and future improvements.
Also, it fixes N+1 queries performed when fetching the list of job
records for both the API endpoints and the web. The number of records is
not significant to notice a difference in performance, but this is a
good optimization to implement from the start.